requirements.txt
Flask==2.3.3
yfinance==0.2.28
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.0
pandas-datareader==0.10.0
matplotlib==3.8.0
plotly==5.17.0
python-dotenv==1.0.0
gunicorn==21.2.0
